backend/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.7
google-api-python-client>=2.116.0
youtube-transcript-api>=1.2.4
httpx>=0.26.0